server/requirements.txt
python-socketio==4.6.0
eventlet==0.25.2